Remove `Generic(F)MapReads`
I did this because the `GenericMapReads` / `GenericFMapReads` typeclasses didn't really represent a new typeclass, and were just used for curried type application of `MapReadsAll` / `MapFReadsAll`.  This commit reworks the type dependency of the underlying typeclasses and adds inference for case classes, obviating the need for the extra typeclasses and unifying it all into one typeclass.